# internal/entity/login.go
package entity

type LoginRequest struct {
	Name     string `json:"name"`
	Password string `json:"password"`
}

type Tokens struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}